Backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.40
# PDF text extraction (PyPDF2 successor; do not add duplicate PyPDF2)
pypdf>=4.0
python-docx>=1.1
gunicorn>=22.0
dj-database-url>=2.2
psycopg2-binary>=2.9
django-cloudinary-storage>=0.3
cloudinary>=1.40
